Added an example to main file
