--- json/test2.json
@@ -1,1 +1,0 @@
-{"ip": "192.168.50.56", "s": "025", "a": "3"}
